+++ R/db.R
@@ -72,6 +72,10 @@
 #' \code{emr_db.init} is the old version of this function which
 #' is now deprecated.
 #'
+#' \code{emr_db.ls} lists all the currently connected databases.
+#'
+#'
+#'
 #' @aliases emr_db.connect emr_db.init_examples
 #' @param db_dirs vector of db directories
 #' @param load_on_demand vector of booleans, same length as db_dirs, if load_on_demand[i] is FALSE, tracks from db_dirs[i] will be pre-loaded, or a single 'TRUE' or 'FALSE' to set \code{load_on_demand} for all the databases. If NULL is passed, \code{load_on_demand} is set to TRUE on all the databases
@@ -139,6 +143,8 @@ emr_db.connect <- function(db_dirs = NULL, load_on_demand = NULL, do_reload = FA
             }
         }
     )
+
+    emr_entries.reload(db_dirs)
 }
 
 #' @export emr_db.init
@@ -161,6 +167,11 @@ emr_db.init <- function(global.dir = NULL, user.dir = NULL, global.load.on.deman
     emr_db.connect(db_dirs = db_dirs, load_on_demand = load_on_demand, do_reload = do.reload)
 }
 
+#' @export
+#' @rdname emr_db.connect
+emr_db.ls <- function() {
+    .naryn$EMR_ROOTS
+}
 
 #' Initialize the examples database
 #'
@@ -231,6 +242,8 @@ emr_db.reload <- function() {
             }
         }
     )
+
+    purrr::walk(emr_db.ls(), emr_entries.reload)
 }
 
 #' Unload all tracks from naryn database