+++ R/zzz.R
@@ -1,9 +1,10 @@
 #' An environment for storing the package global variables
 #'
-#'
+#' @keywords internal
 #' @export
-#' @noRd
 .naryn <- new.env(parent = emptyenv())
+.naryn$entries <- list()
+.naryn$entries_timestamp <- list()
 
 .onLoad <- function(lib, pkg) {
 }